Rename and allow configurable map-form component
