+++ tests/integration/components/map-form-test.js
@@ -3,14 +3,14 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 
-module('Integration | Component | area-map-form', function(hooks) {
+module('Integration | Component | map-form', function(hooks) {
   setupRenderingTest(hooks);
 
   test('it renders', async function(assert) {
     // Set any properties with this.set('myProperty', 'value');
     // Handle any actions with this.set('myAction', function(val) { ... });
 
-    await render(hbs`{{area-map-form}}`);
+    await render(hbs`{{map-form}}`);
 
     assert.equal(this.element.textContent.trim(), '');
   });
